+++ packages/persist_theme/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.3
+
+* Adding Support for ThemeData overrides
+* Adding Default Theme
+* Updated Example
+
 ## 0.0.2
 
 * Fixing Analysis Issues